requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
slowapi==0.1.9
python-multipart==0.0.12

# Document parsing
PyMuPDF==1.24.11          # PDF support
# python-docx==1.1.2      # Uncomment to add DOCX support

# Embeddings
sentence-transformers==3.1.1

# Vector store
faiss-cpu==1.8.0

# LLM
groq==0.11.0
